rename repo name
