--- visualization/render-overlay/VERSION
@@ -1,1 +1,0 @@
-0.1.0-dev2
