--- visualization/render-overlay/build-docker.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-version=$(<VERSION)
-docker build . -t polusai/render-overlay-plugin:${version}
