--- visualization/render-overlay/src/polus/plugins/visualization/render_overlay/__init__.py
@@ -1,2 +1,0 @@
-"""Render Overlay Plugin."""
-__version__ = "0.1.0-dev2"
