--- visualization/render-overlay/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Test for Render Overlay package."""
