+++ visualization/tabular-to-microjson/.bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.1.0-dev2
+current_version = 0.1.0
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+)(?P<dev>\d+))?
@@ -24,4 +24,4 @@ replace = version = "{new_version}"
 
 [bumpversion:file:VERSION]
 
-[bumpversion:file:src/polus/plugins/visualization/render_overlay/__init__.py]
+[bumpversion:file:src/polus/plugins/visualization/tabular_to_microjson/__init__.py]
