+++ visualization/tabular-to-microjson/VERSION
@@ -1,0 +1,1 @@
+0.1.0
