+++ visualization/tabular-to-microjson/build-docker.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+version=$(<VERSION)
+docker build . -t polusai/tabular-to-microjson-plugin:${version}
