+++ visualization/tabular-to-microjson/examples/example_overlay_Point.json
