+++ visualization/tabular-to-microjson/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "polus-plugins-visualization-render-overlay"
-version = "0.1.0-dev2"
+name = "polus-plugins-visualization-tabular-to-microjson"
+version = "0.1.0"
 description = ""
 authors = ["Hamdah Shafqat abbasi <hamdahshafqat.abbasi@nih.gov>"]
 readme = "README.md"
