+++ visualization/tabular-to-microjson/src/polus/plugins/visualization/tabular_to_microjson/__init__.py
@@ -1,0 +1,2 @@
+"""Tabular to microjson package."""
+__version__ = "0.1.0"
