+++ visualization/tabular-to-microjson/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Test for tabular to microjson package."""
